app/requirements.txt
django==3.2
flake8==4.0
python-dotenv==0.21
django-split-settings==1.2
mypy==0.982
pre-commit-hooks==4.3.0
pre_commit==2.20
pydantic==1.10.2
django-debug-toolbar==3.4.0
uwsgi==2.0.21
psycopg2==2.9.5
django-cors-headers==3.13.0
